+++ UnityInventory/Assets/Inventory/Runtime/InventoryConstants.cs
@@ -4,6 +4,5 @@
     {
         public const string CREATE_SO_MENU = "Inventory/";
         public const string CREATE_ITEMS_SM = CREATE_SO_MENU + "Items/";
-        public const string CREATE_ICONS_SM = CREATE_SO_MENU + "Icons/";
     }
 }